Move code to use case
